refactor(dotfiles): src/ をドメイン責務で配線（mod パス・doc リンク）[#515]
純移動（前コミット）を 2018 方式のモジュールパスで繋ぎ、CLI 挙動は不変に保つ。

- main.rs: mod 宣言をドメイン順（共有核 → 配置エンジン → named value →
  hooks → コマンド）へ再編し、入口の読み順を構造で表現する。
- apply.rs をモジュールルート化し copy / compose / generate / gate / strategy を
  子に取り込む。copy / compose → apply の set_mode 借用は super::set_mode の親子
  参照で解消し、set_mode は子からのみ参照されるため private へ絞る。
- locals.rs（ファサード）を新設し store / resolve / inject / prompt を束ねる。
- 子モジュールの可視性は横断参照の有無に合わせる: apply 内部専用の strategy は
  private、他ドメインの rustdoc リンクから参照される他の子は pub(crate)。
- intra-doc リンクを移動後パスへ更新（crate::xxx → crate::apply / locals::*）。
- リンク腐敗の恒久ガード: source.rs の include_dir 曖昧を mod@ で解消し、
  main.rs に #![deny(rustdoc::broken_intra_doc_links)] を追加、CI（rust.yml）に
  cargo doc ステップを足して deny を実効化する。
