+++ src/doctor.rs
@@ -1,13 +1,13 @@
 //! `dotfiles doctor`（雛形・§9.5・§11）: `locals` 宣言がストアに揃っているかを診断する。
 //!
-//! 全単位の `manifest.toml`（[`crate::discover`]）から `locals` を集め、ストア（[`crate::store`]）に
+//! 全単位の `manifest.toml`（[`crate::discover`]）から `locals` を集め、ストア（[`crate::locals::store`]）に
 //! 値が無い名前を報告する。ツール別ロジックは持たず「宣言名がストアに在るか」を見るだけ（§9.5。
 //! git の `git config --includes` 解決スコープに依存しない）。雛形のため依存バイナリ等の他診断は
 //! 後続スライス。未設定があっても**ブロックしない**（exit 0・情報提供）。
 
 use crate::discover::{self, MANIFEST};
+use crate::locals::store::Store;
 use crate::manifest::Manifest;
-use crate::store::Store;
 use std::path::Path;
 
 /// `source` 配下の `locals` 宣言とストアを突き合わせ、未設定を stderr に報告する。
