+++ src/locals/resolve.rs
@@ -1,12 +1,12 @@
 //! apply 時のローカル値解決（§9.3 取得）: 単位の `locals` をストアと突き合わせ、未設定値を
 //! TTY なら対話取得（sensitive は非エコー）、非 TTY なら警告のみで継続する（**ブロックしない**）。
 //!
-//! ここで解決できた「名前→値」だけを [`crate::inject`] が placeholder 置換に使う。未解決の名前は
+//! ここで解決できた「名前→値」だけを [`crate::locals::inject`] が placeholder 置換に使う。未解決の名前は
 //! マップに入らないため、配置ファイルでは `@@name@@` が literal のまま残り doctor が検出する（§9.1）。
 
+use super::store::Store;
+use super::{inject, prompt};
 use crate::manifest::Manifest;
-use crate::store::Store;
-use crate::{inject, prompt};
 use std::collections::BTreeMap;
 
 /// 単位の `locals` を解決し、注入用の「名前→値」マップ（解決できた名前のみ）を返す。
