+++ src/main.rs
@@ -8,35 +8,38 @@
 //! `manifest.toml` に従って配置する。配置は **2軸**
 //! （生成方式 `kind`=copy/generate × 合成 `strategy`=concat/json-shallow）＋条件付き overlay
 //! （`when` gate）で捉える（設計書 §5 / §5.5）。copy はツリー配置、generate / overlay 明示は
-//! ファイル合成（[`compose`]）を通り、トップレベル `when`（`deps` / `os`）はユニット単位 gate（[`gate`]）。配置の直前に
-//! `locals`（named value）を解決・注入する（[`resolve`] / [`inject`] / [`store`] / [`prompt`]、§9）。
+//! ファイル合成（[`crate::apply::compose`]）を通り、トップレベル `when`（`deps` / `os`）はユニット単位 gate（[`crate::apply::gate`]）。配置の直前に
+//! `locals`（named value）を解決・注入する（[`crate::locals::resolve`] / [`crate::locals::inject`] / [`crate::locals::store`] / [`crate::locals::prompt`]、§9）。
 //! 配置後は `hooks`（onchange フック）をユニットのソースハッシュ変化時だけ実行する
 //! （[`hooks`] / [`onchange`]、§13）。`apply` は配置＋フック、`list` は配置先一覧、`secret set` は
 //! named value 設定、`color sample` は ANSI 確認表（旧 `crates/color` を吸収、§10）、`doctor` は診断（雛形）。
+
+#![deny(rustdoc::broken_intra_doc_links)]
 
 use clap::{Parser, Subcommand};
 use std::ffi::OsString;
 use std::path::{Path, PathBuf};
 
+// 共有核（葉。多くの群が片方向で依存する契約・基盤）。
+mod discover; // §6.3 走査（apply / list / doctor 共有）
+mod manifest; // §6 manifest.toml スキーマ
+mod source; // §8 ソース二段構え
+
+// 配置エンジン（§5 / §5.5）。子モジュール（copy / compose / generate / strategy / gate）は apply.rs が束ねる。
 mod apply;
-mod color;
-mod compose;
-mod copy;
-mod discover;
-mod doctor;
-mod gate;
-mod generate;
+
+// named value（§9）。子モジュール（store / resolve / inject / prompt）は locals.rs が束ねる。
+mod locals;
+
+// onchange フック（§13。2 ファイルなので直下据え置き）。
 mod hooks;
-mod inject;
-mod list;
-mod manifest;
 mod onchange;
-mod prompt;
-mod resolve;
+
+// 単独ビュー / コマンド。
+mod color;
+mod doctor;
+mod list;
 mod secret;
-mod source;
-mod store;
-mod strategy;
 
 /// toshiki670/dotfiles 本体（core）。
 #[derive(Parser)]
