+++ src/onchange.rs
@@ -6,7 +6,7 @@
 //!   ハッシュを採る（mtime は touch/clone で揺れるため内容ハッシュ）。chezmoi の
 //!   `run_onchange_*` が埋め込みソースの `sha256sum` を比較していたのと同じ役割。
 //! - **状態の永続化**（[`State`]）: フックごとの前回ハッシュを `~/.config/dotfiles/hooks.toml` に
-//!   持つ。秘匿値ではない（[`crate::store`] と違い 0600 不要）ため平文で書き、破損時は warn して
+//!   持つ。秘匿値ではない（[`crate::locals::store`] と違い 0600 不要）ため平文で書き、破損時は warn して
 //!   空（全フック再実行）扱いにする ― 再実行は冪等なので、致命的エラーで apply を止めるより
 //!   作り直す方が安全（disposable な状態）。
